Arrow Functions
created a variable called divide and assigned it to an arrow function that divides 2000 by 100. created a second variable called square that was assigned to an arrow function that takes in one parameter and multiplies it by itself. created a third variable  and assigned it to an arrow functio that takes in two parameters that get added together